Fixed Psalm issues for level 5
Psalm actually found an issue!
Here an instance of an interface was injected into the controller via Laravels
DI System, but the methods called on the injected instance were actually
ones only available of a specific Implementation of that interface.

We fix this by injecting an instance of the specific implementation instead of
the interface.