Reduce cyclomatic complexity
We introduce a `UserService`, which encapsulates our business
logic and is separate from our HTTP layer. We also introduce an
`UpdateUserRequest`, which acts as a type-safe container and
validator for the user input via HTTP.

Now the controller is only responsible for delegating tasks to
the more specialized components and the cyclomatic complexity is
below our (artificially small) threshold!